+++ src/core/crypto/ed25519.ts
@@ -1,7 +1,7 @@
 // Copyright © Aptos Foundation
 // SPDX-License-Identifier: Apache-2.0
 
-import nacl from "tweetnacl";
+import { ed25519 } from "@noble/curves/ed25519";
 import { Deserializer } from "../../bcs/deserializer";
 import { Serializable, Serializer } from "../../bcs/serializer";
 import { AuthenticationKey } from "../authenticationKey";
@@ -16,6 +16,14 @@ import { PrivateKey } from "./privateKey";
 import { AccountPublicKey, VerifySignatureArgs } from "./publicKey";
 import { Signature } from "./signature";
 import { convertSigningMessage } from "./utils";
+
+/**
+ * L is the value that greater than or equal to will produce a non-canonical signature, and must be rejected
+ */
+const L: number[] = [
+  0xed, 0xd3, 0xf5, 0x5c, 0x1a, 0x63, 0x12, 0x58, 0xd6, 0x9c, 0xf7, 0xa2, 0xde, 0xf9, 0xde, 0x14, 0x00, 0x00, 0x00,
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x10,
+];
 
 /**
  * Represents the public key of an Ed25519 key pair.
@@ -69,7 +77,12 @@ export class Ed25519PublicKey extends AccountPublicKey {
     const messageBytes = Hex.fromHexInput(messageToVerify).toUint8Array();
     const signatureBytes = signature.toUint8Array();
     const publicKeyBytes = this.key.toUint8Array();
-    return nacl.sign.detached.verify(messageBytes, signatureBytes, publicKeyBytes);
+    // Also verify malleability
+    if (!signature.isCanonicalSignature()) {
+      return false;
+    }
+
+    return ed25519.verify(signatureBytes, messageBytes, publicKeyBytes);
   }
 
   authKey(): AuthenticationKey {
@@ -130,7 +143,7 @@ export class Ed25519PrivateKey extends Serializable implements PrivateKey {
    * The Ed25519 signing key
    * @private
    */
-  private readonly signingKeyPair: nacl.SignKeyPair;
+  private readonly signingKey: Hex;
 
   // region Constructors
 
@@ -148,7 +161,7 @@ export class Ed25519PrivateKey extends Serializable implements PrivateKey {
     }
 
     // Create keyPair from Private key in Uint8Array format
-    this.signingKeyPair = nacl.sign.keyPair.fromSeed(privateKeyHex.toUint8Array().slice(0, Ed25519PrivateKey.LENGTH));
+    this.signingKey = privateKeyHex;
   }
 
   /**
@@ -157,8 +170,8 @@ export class Ed25519PrivateKey extends Serializable implements PrivateKey {
    * @returns Ed25519PrivateKey
    */
   static generate(): Ed25519PrivateKey {
-    const keyPair = nacl.sign.keyPair();
-    return new Ed25519PrivateKey(keyPair.secretKey.slice(0, Ed25519PrivateKey.LENGTH));
+    const keyPair = ed25519.utils.randomPrivateKey();
+    return new Ed25519PrivateKey(keyPair);
   }
 
   /**
@@ -189,7 +202,7 @@ export class Ed25519PrivateKey extends Serializable implements PrivateKey {
    * @returns Ed25519PublicKey
    */
   publicKey(): Ed25519PublicKey {
-    const bytes = this.signingKeyPair.publicKey;
+    const bytes = ed25519.getPublicKey(this.signingKey.toUint8Array());
     return new Ed25519PublicKey(bytes);
   }
 
@@ -202,7 +215,7 @@ export class Ed25519PrivateKey extends Serializable implements PrivateKey {
   sign(message: HexInput): Ed25519Signature {
     const messageToSign = convertSigningMessage(message);
     const messageBytes = Hex.fromHexInput(messageToSign).toUint8Array();
-    const signatureBytes = nacl.sign.detached(messageBytes, this.signingKeyPair.secretKey);
+    const signatureBytes = ed25519.sign(messageBytes, this.signingKey.toUint8Array());
     return new Ed25519Signature(signatureBytes);
   }
 
@@ -212,7 +225,7 @@ export class Ed25519PrivateKey extends Serializable implements PrivateKey {
    * @returns Uint8Array representation of the private key
    */
   toUint8Array(): Uint8Array {
-    return this.signingKeyPair.secretKey.slice(0, Ed25519PrivateKey.LENGTH);
+    return this.signingKey.toUint8Array();
   }
 
   /**
@@ -221,7 +234,7 @@ export class Ed25519PrivateKey extends Serializable implements PrivateKey {
    * @returns string representation of the private key
    */
   toString(): string {
-    return Hex.fromHexInput(this.toUint8Array()).toString();
+    return this.signingKey.toString();
   }
 
   // endregion
@@ -292,6 +305,27 @@ export class Ed25519Signature extends Signature {
   static deserialize(deserializer: Deserializer): Ed25519Signature {
     const bytes = deserializer.deserializeBytes();
     return new Ed25519Signature(bytes);
+  }
+
+  /**
+   * Checks if an ED25519 signature is non-canonical.
+   *
+   * Comes from Aptos Core
+   * https://github.com/aptos-labs/aptos-core/blob/main/crates/aptos-crypto/src/ed25519/ed25519_sigs.rs#L47-L85
+   */
+  isCanonicalSignature(): boolean {
+    const s = this.toUint8Array().slice(32);
+
+    for (let i = s.length - 1; i >= 0; i -= 1) {
+      if (s[i] < L[i]) {
+        return true;
+      }
+      if (s[i] > L[i]) {
+        return false;
+      }
+    }
+    // As this stage S == L which implies a non-canonical S.
+    return false;
   }
 
   // endregion