+++ tests/e2e/api/transaction.test.ts
@@ -4,6 +4,7 @@
 import { Account, TransactionResponse, U64 } from "../../../src";
 import { FUND_AMOUNT } from "../../unit/helper";
 import { getAptosClient } from "../helper";
+import { longWaitForTransaction } from "../../../src/internal/transaction";
 
 // use it here since all tests use the same configuration
 const { aptos } = getAptosClient();
@@ -77,6 +78,38 @@ describe("transaction api", () => {
         ledgerVersion: Number(txn.version),
       });
       expect(transaction).toStrictEqual(txn);
+    });
+
+    test("it queries for transactions by hash", async () => {
+      const transaction = await aptos.getTransactionByHash({
+        transactionHash: txn.hash,
+      });
+      expect(transaction).toStrictEqual(txn);
+    });
+  });
+
+  describe("long poll", () => {
+    let txn: TransactionResponse;
+    beforeAll(async () => {
+      const senderAccount = Account.generate();
+      await aptos.fundAccount({ accountAddress: senderAccount.accountAddress, amount: FUND_AMOUNT });
+      const bob = Account.generate();
+      const rawTxn = await aptos.transaction.build.simple({
+        sender: senderAccount.accountAddress,
+        data: {
+          function: "0x1::aptos_account::transfer",
+          functionArguments: [bob.accountAddress, 10],
+        },
+      });
+      const authenticator = aptos.transaction.sign({
+        signer: senderAccount,
+        transaction: rawTxn,
+      });
+      const response = await aptos.transaction.submit.simple({
+        transaction: rawTxn,
+        senderAuthenticator: authenticator,
+      });
+      txn = await longWaitForTransaction({ aptosConfig: aptos.config, transactionHash: response.hash });
     });
 
     test("it queries for transactions by hash", async () => {